@@ -1,44 +1,56 @@
 {
   "name": "StandardNotes",
-  "version": "3.0.17",
-  "versionIOS": "3.0.17",
-  "versionAndroid": "3.0.17",
+  "version": "3.0.18",
+  "versionIOS": "3.0.18",
+  "versionAndroid": "3.0.18",
   "license": "AGPL-3.0-or-later",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "android": "cd ./android && ./gradlew app:assembleDebug && ./gradlew installDebug",
+    "init": "git submodule update --init && npm install && cd ios && pod install",
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
+    "start": "react-native start",
+    "test": "jest",
+    "lint": "eslint .",
     "clear-cache": "watchman watch-del-all && rm -rf $TMPDIR/react-native-packager-cache-* && rm -rf $TMPDIR/metro-bundler-cache-*",
     "build-android": "cd android && ./gradlew bundleRelease"
   },
   "dependencies": {
-    "@react-native-community/async-storage": "1.4.0",
+    "@react-native-community/async-storage": "1.6.3",
     "base-64": "^0.1.0",
-    "bugsnag-react-native": "^2.12.6",
+    "bugsnag-react-native": "^2.23.2",
     "immutable": "^3.8.2",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "react": "16.8.3",
-    "react-native": "0.59.10",
-    "react-native-file-viewer": "^1.0.10",
-    "react-native-fs": "^2.13.3",
-    "react-native-gesture-handler": "^1.0.12",
-    "react-native-keychain": "^3.1.3",
-    "react-native-store-review": "^0.1.3",
+    "react": "16.12.0",
+    "react-native": "0.61.5",
+    "react-native-file-viewer": "^2.0.0",
+    "react-native-fs": "^2.16.2",
+    "react-native-gesture-handler": "^1.5.2",
+    "react-native-keychain": "^4.0.1",
+    "react-native-reanimated": "^1.4.0",
+    "react-native-screens": "^1.0.0-alpha.23",
+    "react-native-store-review": "^0.1.5",
     "react-native-vector-icons": "6.6.0",
-    "react-native-webview": "5.8.1",
-    "react-navigation": "^3.13.0",
+    "react-native-webview": "https://github.com/standardnotes/react-native-webview.git#92aac04e0f4e1e27edbce9e053436c6777a6f34a",
+    "react-navigation": "^4.0.10",
+    "react-navigation-drawer": "^2.3.3",
     "react-navigation-header-buttons": "^2.1.1",
-    "regenerator": "^0.13.3",
-    "snjs": "0.2.8",
-    "standard-file-js": "0.3.69",
-    "stacktrace-parser": "0.1.4"
+    "react-navigation-stack": "^1.10.3",
+    "regenerator": "^0.14.2",
+    "snjs": "0.2.9",
+    "stacktrace-parser": "0.1.8",
+    "standard-file-js": "0.3.71"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/runtime": "^7.4.5",
-    "metro-react-native-babel-preset": "^0.54.1",
-    "react-test-renderer": "16.8.3"
+    "@babel/core": "^7.7.4",
+    "@babel/runtime": "^7.7.4",
+    "@react-native-community/eslint-config": "^0.0.5",
+    "babel-jest": "^24.9.0",
+    "eslint": "^6.7.2",
+    "jest": "^24.9.0",
+    "metro-react-native-babel-preset": "^0.57.0",
+    "react-test-renderer": "16.12.0"
   },
   "jest": {
     "preset": "react-native"
