@@ -58,35 +58,38 @@
 }
 
 - (void)configurePinning {
-  NSDictionary *trustKitConfig =
-  @{
-    kTSKSwizzleNetworkDelegates: @YES,
+  if(@available(iOS 10, *)) {
+    NSDictionary *trustKitConfig =
+    @{
+      kTSKSwizzleNetworkDelegates: @YES,
+
+      // The list of domains we want to pin and their configuration
+      kTSKPinnedDomains: @{
+        @"standardnotes.org" : @{
+          kTSKIncludeSubdomains:@YES,
+
+          kTSKEnforcePinning:@YES,
+
+          // Send reports for pin validation failures so we can track them
+          kTSKReportUris: @[@"https://standard.report-uri.com/r/d/hpkp/reportOnly"],
+
+          // The pinned public keys' Subject Public Key Info hashes
+          kTSKPublicKeyHashes : @[
+            @"Vjs8r4z+80wjNcr1YKepWQboSIRi63WsWXhIMN+eWys=",
+            @"C5+lpZ7tcVwmwQIMcRtPbsQtWLABXhQzejna0wHFr8M=",
+            @"YLh1dUR9y6Kja30RrAn7JKnbQG/uEtLMkBgFF2Fuihg=",
+            @"sRHdihwgkaib1P1gxX8HFszlD+7/gTfNvuAybgLPNis=",
+            @"++MBgDH5WGvL9Bcn5Be30cRcL0f5O+NyoXuWtQdX1aI=",
+            @"f0KW/FtqTjs108NpYj42SrGvOB2PpxIVM8nWxjPqJGE=",
+            @"NqvDJlas/GRcYbcWE8S/IceH9cq77kg0jVhZeAPXq8k=",
+            @"9+ze1cZgR9KO1kZrVDxA4HQ6voHRCSVNz4RdTCx4U8U="
+          ],
+        },
+      }
+    };
     
-    // The list of domains we want to pin and their configuration
-    kTSKPinnedDomains: @{
-      @"standardnotes.org" : @{
-        kTSKIncludeSubdomains:@YES,
-        
-        kTSKEnforcePinning:@YES,
-      
-        // Send reports for pin validation failures so we can track them
-        kTSKReportUris: @[@"https://standard.report-uri.com/r/d/hpkp/reportOnly"],
-        
-        // The pinned public keys' Subject Public Key Info hashes
-        kTSKPublicKeyHashes : @[
-          @"Vjs8r4z+80wjNcr1YKepWQboSIRi63WsWXhIMN+eWys=",
-          @"C5+lpZ7tcVwmwQIMcRtPbsQtWLABXhQzejna0wHFr8M=",
-          @"YLh1dUR9y6Kja30RrAn7JKnbQG/uEtLMkBgFF2Fuihg=",
-          @"sRHdihwgkaib1P1gxX8HFszlD+7/gTfNvuAybgLPNis=",
-          @"++MBgDH5WGvL9Bcn5Be30cRcL0f5O+NyoXuWtQdX1aI=",
-          @"f0KW/FtqTjs108NpYj42SrGvOB2PpxIVM8nWxjPqJGE=",
-          @"NqvDJlas/GRcYbcWE8S/IceH9cq77kg0jVhZeAPXq8k=",
-          @"9+ze1cZgR9KO1kZrVDxA4HQ6voHRCSVNz4RdTCx4U8U="
-        ],
-      },
-    }
-  };
-  [TrustKit initSharedInstanceWithConfiguration:trustKitConfig];
+    [TrustKit initSharedInstanceWithConfiguration:trustKitConfig];
+  }
 }
 
 - (NSURL *)sourceURLForBridge:(RCTBridge *)bridge
