@@ -1,5 +1,5 @@
 import React, { Component, Fragment } from 'react';
-import {Keyboard} from 'react-native';
+import { Keyboard } from 'react-native';
 import Abstract from "@Screens/Abstract"
 
 export default class AbstractSideMenu extends Abstract {
@@ -9,40 +9,26 @@ export default class AbstractSideMenu extends Abstract {
       We had some performance issues with this component rendering too many times when
       navigating to unrelated routes, like pushing Compose. It would render 6 times or so,
       causing slowdown. We only want to render if there's a difference in drawer related properties.
-      I've found that when the drawer is about to appear, one of these params will change: "isDrawerIdle", "isDrawerOpen", "isTransitioning"
     */
-    let newNavigationState = nextProps.navigation.state;
-    let currentNavigationState = this.props.navigation.state;
 
-    // When navigating to Compose from Notes, we don't want this to render. I've found that the drawerMovementDirection is
-    // set to 'closing' in this circumstance.
-    if(newNavigationState.drawerMovementDirection == "closing" || currentNavigationState.drawerMovementDirection == "closing") {
-      return false;
-    }
+    const isDrawerOpen = this.props.navigation.state.isDrawerOpen;
+    const isDrawerGoingToBeOpen = nextProps.navigation.state.isDrawerOpen;
 
-    let isSame = Abstract.IsShallowEqual(
-      newNavigationState,
-      currentNavigationState,
-      ["isDrawerIdle", "isDrawerOpen", "isTransitioning"]
-    );
-
-    let isSameIdleState = newNavigationState.isDrawerIdle == currentNavigationState.isDrawerIdle
-
-    // These are just the sort of chance configuration of variables that indicate that a drawer will focus, and have it called only once
-    if(!isSame && !newNavigationState.isDrawerIdle && !isSameIdleState) {
+    // When the drawer is opening, we want to be sure to dismiss the keyboard if it's up
+    if(!isDrawerOpen && isDrawerGoingToBeOpen) {
       this.psuedo_willFocus();
     }
-    return !isSame;
+
+    // We only need this to render when the drawer is opening, not when closing.
+    return !isDrawerOpen && isDrawerGoingToBeOpen;
   }
 
   psuedo_willFocus() {
     // componentWillFocus is not called for drawer components when they are about to appear
     // instead, we piggyback on the logic in shouldComponentUpdate above to determine
     // if navigation state is about to change, and if so, we call this.
-
     Keyboard.dismiss();
 
     this.handler && this.handler.onKeyboardDismiss && this.handler.onKeyboardDismiss();
   }
-
 }
